Criada Rede Neural do Piercing Line e Dark Cloud

+++ src/main/java/br/iesb/indicador_analise_grafica_enum/Perfil.java
@@ -2,19 +2,19 @@ package br.iesb.indicador_analise_grafica_enum;
 
 public enum Perfil {
 	
-	ARROJADO("Arrojado", 60),
-	MODERADO("Moderado", 70),
-	CONSERVADOR("Conservador", 80);
+	ARROJADO("Arrojado", 60.0),
+	MODERADO("Moderado", 70.0),
+	CONSERVADOR("Conservador", 80.0);
 	
 	String definicao;
-	int id;
+	Double id;
 	
-	Perfil(String definicao, int id){
+	Perfil(String definicao, Double id){
 		this.definicao = definicao;
 		this.id = id;
 	}
 	
-	public int getValor() {
+	public Double getValor() {
 		return id;
 	}
